Fix specifying the number of octects in a character
Binutils 2.44 is tightening up some of the code in the assembler.

As a result TF-A fails to build when using it because the plat_macro.S
file is creating a string section without specifying the number of
octets that make up a single character in the string.  The assembler
used to allow a default of 1 string character = 1 octet, but with
the update it is now up to the coder to provide an explicit string
size.

The good news is that the change is very easy. We just have to change
instances of .rodata.gic_reg_name in the various plat_macros.S from:

.section .rodata.gic_reg_name, "aS"

  to:

.section .rodata.gic_reg_name, "aS", 1

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>
Change-Id: Ia5b23b7bee5f83cb40c17102112ecc1e40cc5724